+++ software/iob-gpio.h
@@ -26,6 +26,6 @@ void trigger_output(uint8_t value);
 
 void gpio_set_cathode_output(uint32_t);
 
-void gpio_set_anode_output(uint32_t);
+void gpio_set_anode_output(uint8_t);
